python-pipeline-sandbox/requirements.txt
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
requests>=2.25.0
trafilatura>=1.6.0
readability-lxml>=0.8.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
supabase>=2.6.0